+++ PowerShell.exe/Not-current-version/.test.ps1
@@ -14,6 +14,8 @@ task Not-current-version -If $IsV3Win {
 task Not-same-exe -If $IsV6Core {
 	($r = ./Not-same-exe.ps1)
 	equals $r.Count 2
-	equals $r[0] "$PSVersion" # fixed in v6-beta.7
+	# fixed in v6-beta.7
+	# changed in v6-beta.9
+	assert ($r[0] -ne "$PSVersion")
 	equals $r[1] "$PSVersion"
 }