+++ PowerShell.exe/Version-parameter/v6-ignores-stuff-after-Version.ps1
@@ -1,0 +1,3 @@
+
+# v6-beta.9 - anything after -Version is silently ignored
+pwsh -NoProfile -Version 2 bar bar